fixed services section to add what client wanted, added links to the social media buttons
