+++ src/components/Atoms/Icons/icons.css
@@ -1,3 +1,3 @@
 .socialIcon {
-    transform: scale(1.5);
+    transform: scale(2.5);
 }
