+++ src/components/Atoms/Pictures/Pictures.css
@@ -5,11 +5,14 @@
 
 .logoFooterContainer {
     height: 100%;
+    display: flex;
+    align-items: center;
 }
 
 .logoFooter {
     display: block;
-    height: 100%;
+    height: 50%;
+    width: auto;
 }
 
 .heroImageContainer {
