--- src/components/Atoms/Text/HomePage/ServicesText/Facials.js
@@ -1,13 +1,0 @@
-export default function Facials() {
-    return (
-        <div className="serviceContainer">
-            <h1 className="serviceName">Facial Treatments</h1>
-            <p className="serviceText">
-                Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do
-                eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut
-                enim ad minim veniam, quis nostrud exercitation ullamco laboris
-                nisi ut aliquip ex ea commodo consequat.
-            </p>
-        </div>
-    );
-}
