+++ src/components/Atoms/Text/HomePage/ServicesText/VirtualService.js
@@ -1,0 +1,15 @@
+import "./servicesText.css";
+
+export default function VirtualService() {
+    return (
+        <div className="serviceContainer">
+            <h1 className="serviceName">Virtual Services</h1>
+            <p className="serviceText">
+                Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do
+                eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut
+                enim ad minim veniam, quis nostrud exercitation ullamco laboris
+                nisi ut aliquip ex ea commodo consequat.
+            </p>
+        </div>
+    );
+}
