+++ src/components/Molecules/ServicesBox/servicesBox.css
@@ -2,6 +2,7 @@
     display: flex;
     flex-direction: row;
     justify-content: space-evenly;
+    flex-wrap: wrap;
     width: 100%;
     padding: 30px;
     margin: 20px 0 20px 0;
