+++ src/components/Molecules/SocialLinks/Socials.js
@@ -1,6 +1,5 @@
 import "./socials.css";
 import FacebookIcon from "../../Atoms/Icons/Facebook";
-import TwitterIcon from "../../Atoms/Icons/Twitter";
 import InstagramIcon from "../../Atoms/Icons/Instagram";
 
 export default function Socials() {
@@ -9,7 +8,6 @@ export default function Socials() {
             <p>Join Me On My Socials!</p>
             <div className="socialsContainer">
                 <FacebookIcon />
-                <TwitterIcon />
                 <InstagramIcon />
             </div>
         </div>
