+++ src/components/Organisms/HomePage/Footer/footer.css
@@ -1,11 +1,10 @@
 .footerContainer {
     display: flex;
     flex-direction: row;
-    justify-content: space-around;
+    justify-content: space-between;
     align-items: center;
     width: 100%;
     height: 300px;
     z-index: 998;
     background: #202020;
-    padding: 30px;
 }
