+++ src/components/Organisms/HomePage/Header/Header.css
@@ -1,5 +1,5 @@
 .heroContainer {
-    background-image: url("../../../../Assets/Pictures/heroHome2.avif");
+    background-image: url("../../../../Assets/Pictures/HeroImage/heroHome2.avif");
     background-repeat: no-repeat;
     background-size: cover;
     width: 100%;
